+++ src/pages/RegisterPage/RegisterPage.style.ts
